backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy>=2.0
alembic>=1.13
pydantic>=2.6
jsonschema>=4.21
python-dotenv>=1.0
httpx>=0.27
pytest>=8.1
pytest-cov>=5.0
ruff>=0.4
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
passlib[bcrypt]>=1.7
twilio>=9.0
sentry-sdk[fastapi]>=2.0
prometheus-fastapi-instrumentator>=2.0
